Refactors placement of redux-reducer/action/types.
