+++ src/users.action.js
@@ -1,0 +1,9 @@
+import RECEIVE_DATA from "./users.types";
+
+
+export function receiveDataAction (users) {
+  return {
+    type: RECEIVE_DATA,
+    users,
+  }
+}
