+++ src/users.reducer.js
@@ -1,0 +1,10 @@
+import RECEIVE_DATA from "./users.types";
+
+export function users(state= [], action){
+    switch(action.type){
+        case RECEIVE_DATA:
+            return action.users
+        default:
+            return state
+    }
+}
