+++ src/users.types.js
@@ -1,0 +1,3 @@
+const RECEIVE_DATA = "RECEIVE_DATA";
+
+export default RECEIVE_DATA
